actualizacion

+++ database/migrations/2024_01_11_020154_create_inversion_publica_efectiva_sectors_table.php
@@ -19,7 +19,7 @@ class CreateInversionPublicaEfectivaSectorsTable extends Migration
             $table->foreign('InversionPublicaEfectiva_id')->references('id')->on('InversionPublicaEfectiva')->onDelete('cascade');
             $table->string('sector');
             $table->integer('inversionD');
-            $table->integer('inversionP');
+            $table->decimal('inversionP', 8, 2);
             $table->timestamps();
         });
     }
